Backend/requirements.txt
Flask==3.0.0
cryptography==41.0.7
google-cloud-kms==2.21.3
requests==2.31.0
Pillow
gunicorn==21.2.0
tzdata
